+++ index.html
@@ -7,6 +7,7 @@
     <title>Muhammed's Portfolio</title>
     <link rel="stylesheet" href="style.css" />
     <script src="scripts/app.js" defer></script>
+    <script src="scripts/works.js" type="module"></script>
   </head>
   <body>
     <header class="header">
