+++ scripts/db.js
@@ -1,0 +1,49 @@
+
+const data = [
+    {
+      name: ["Tonic"],
+      description: [
+        "A daily selection of privately personalized reads; no accounts or sign-ups required.",
+      ],
+      featured: [{ feature1: "CANOPY", feature2: "Back End Dev", year: 2023 }],
+      image: [{ imageSrc: "../assets/m-project-1.png", imageAlt: "project-4" }],
+      technologies: [{ teck1: "html", teck2: "css", teck3: "javascript" }],
+      liveLink: [],
+      sourceLink: [],
+    },
+    {
+      name: ["Multi-Post Stories"],
+      description: [
+        "Experimental content creation feature that allows users to add to an existing story over the course of a day without spamming their friends.",
+      ],
+      featured: [{ feature1: "CANOPY", feature2: "Back End Dev", year: 2023 }],
+      image: [{ imageSrc: "../assets/m-project-2.png", imageAlt: "project-4" }],
+      technologies: [{ teck1: "html", teck2: "css", teck3: "javascript" }],
+      liveLink: [],
+      sourceLink: [],
+    },
+    {
+      name: ["Facebook 360"],
+      description: [
+        "Exploring the future of media in Facebook's first Virtual Reality app; a place to discover and enjoy 360 photos and videos on Gear VR.",
+      ],
+      featured: [{ feature1: "CANOPY", feature2: "Back End Dev", year: 2023 }],
+      image: [{ imageSrc: "../assets/m-project-3.png", imageAlt: "project-4" }],
+      technologies: [{ teck1: "html", teck2: "css", teck3: "javascript" }],
+      liveLink: [],
+      sourceLink: [],
+    },
+    {
+      name: ["Uber Navigation"],
+      description: [
+        "A smart assistant to make driving more safe, efficient, and fun by unlocking your most expensive computer: your car.",
+      ],
+      featured: [{ feature1: "CANOPY", feature2: "Back End Dev", year: 2023 }],
+      image: [{ imageSrc: "../assets/m-project-4.png", imageAlt: "project-4" }],
+      technologies: [{ teck1: "html", teck2: "css", teck3: "javascript" }],
+      liveLink: [],
+      sourceLink: [],
+    },
+  ];
+
+  export default data;
